pyproject.toml
[tool.poetry]
name = "example-refactor-mail"
version = "0.1.0"
description = "Example refactor send mail to microservice process"
authors = ["Jônatas Oliveira <contato@jonatasoliveira.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
flask-restx = "^0.2.0"
flask-sqlalchemy = "^2.4.1"
flask-migrate = "^2.5.3"
pytest = "^5.4.1"
flask-marshmallow = "^0.11.0"
python-decouple = "^3.3"
requests = "^2.23.0"
psycopg2-binary = "^2.8.5"

[tool.poetry.dev-dependencies]
jedi = "^0.16.0"
neovim = "^0.3.1"
ipdb = "^0.13.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
